% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaModelCommons.R
\name{createGOFplot}
\alias{createGOFplot}
\title{Create a Plot of Observed and Predicted Values}
\usage{
createGOFplot(dataObject, predictions = NULL, title = NULL, textsize = 20)
}
\arguments{
\item{dataObject}{a data.frame object with the observations}

\item{predictions}{a data.frame object with the predictions}

\item{title}{a character string}

\item{textsize}{the font size (20 by default)}
}
\description{
The predicted values are optional.
}
